+++ plone/app/contentrules/tests/test_rule_assignment_mapping.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 import transaction
 from zope.component import getUtility