+++ plone/app/contentrules/tests/test_setup.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 from zope.lifecycleevent.interfaces import IObjectModifiedEvent
 from zope.lifecycleevent.interfaces import IObjectAddedEvent
 from zope.lifecycleevent.interfaces import IObjectRemovedEvent