+++ plone/app/contentrules/tests/utils.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 import doctest
 
 # Standard options for DocTests